Reorganized menu buttons
