+++ src/Player.py
@@ -1,0 +1,4 @@
+# TODO: Implement class Player
+
+class Player:
+    pass
